pyproject.toml
[project]
name = "cashflow-project"
version = "0.1.0"
description = "Django cashflow management application"
authors = [
    {name = "Vladislav", email = "vladosmen29@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Django>=4.2,<5.0",
    "django-filter>=23.0",
    "django-widget-tweaks>=1.4.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]